+++ .docker/Dockerfile
@@ -73,7 +73,6 @@ COPY ghost/admin/package.json ghost/admin/package.json
 COPY ghost/announcement-bar-settings/package.json ghost/announcement-bar-settings/package.json
 COPY ghost/api-framework/package.json ghost/api-framework/package.json
 COPY ghost/api-version-compatibility-service/package.json ghost/api-version-compatibility-service/package.json
-COPY ghost/captcha-service/package.json ghost/captcha-service/package.json
 COPY ghost/constants/package.json ghost/constants/package.json
 COPY ghost/core/package.json ghost/core/package.json
 COPY ghost/custom-fonts/package.json ghost/custom-fonts/package.json
