+++ compose.yml
@@ -38,7 +38,6 @@ services:
       - node_modules_ghost_announcement-bar-settings:/home/ghost/ghost/announcement-bar-settings/node_modules:delegated
       - node_modules_ghost_api-framework:/home/ghost/ghost/api-framework/node_modules:delegated
       - node_modules_ghost_api-version-compatibility-service:/home/ghost/ghost/api-version-compatibility-service/node_modules:delegated
-      - node_modules_ghost_captcha-service:/home/ghost/ghost/captcha-service/node_modules:delegated
       - node_modules_ghost_constants:/home/ghost/ghost/constants/node_modules:delegated
       - node_modules_ghost_core:/home/ghost/ghost/core/node_modules:delegated
       - node_modules_ghost_custom-fonts:/home/ghost/ghost/custom-fonts/node_modules:delegated
@@ -192,7 +191,6 @@ volumes:
   node_modules_ghost_announcement-bar-settings: {}
   node_modules_ghost_api-framework: {}
   node_modules_ghost_api-version-compatibility-service: {}
-  node_modules_ghost_captcha-service: {}
   node_modules_ghost_constants: {}
   node_modules_ghost_core: {}
   node_modules_ghost_custom-fonts: {}
