+++ ghost/core/core/server/services/members/api.js
@@ -18,7 +18,7 @@ const tiersService = require('../tiers');
 const newslettersService = require('../newsletters');
 const memberAttributionService = require('../member-attribution');
 const emailSuppressionList = require('../email-suppression-list');
-const CaptchaService = require('@tryghost/captcha-service');
+const CaptchaService = require('./CaptchaService');
 const {t} = require('../i18n');
 const sentry = require('../../../shared/sentry');
 const sharedConfig = require('../../../shared/config');
