+++ ghost/core/package.json
@@ -71,7 +71,6 @@
     "@tryghost/api-framework": "0.0.0",
     "@tryghost/api-version-compatibility-service": "0.0.0",
     "@tryghost/bookshelf-plugins": "0.6.25",
-    "@tryghost/captcha-service": "0.0.0",
     "@tryghost/color-utils": "0.2.2",
     "@tryghost/config-url-helpers": "1.0.12",
     "@tryghost/constants": "0.0.0",
@@ -181,6 +180,7 @@
     "got": "11.8.6",
     "gscan": "4.47.0",
     "handlebars": "4.7.8",
+    "hcaptcha": "0.2.0",
     "human-number": "2.0.4",
     "iconv-lite": "0.6.3",
     "image-size": "1.2.0",
